Read logging config from separate file
* Move all settings helper-functions to `easydmp.site.settings`
* Allow easy overriding of loglevels, by the name of a handler or logger
* Include the default production logging setup in a new module

This incidentally solves a problem on Django 3.1 triggered by running
`logging.config.dictConfig` multiple times or logging something before
logging is set up. This leads to an ImproperlyConfigured exception when
loading settings: SECRET_KEY cannot be found.

(Why this does not happen on Django 3.0 or older is a bit of a mystery.
After a lengthy debugging/yakshave session, working around the problem
seemed the best and most permanent solution.)
